feat: share source content parser
move project and writing source-content parsing into @anipotts/content so apps/admin only owns the raw markdown import boundary. this keeps /content read-only while aligning shared content logic with packages/content.

verify: pnpm --filter @anipotts/content build

verify: pnpm turbo typecheck --filter=@anipotts/admin...

verify: pnpm turbo build --filter=@anipotts/admin...

verify: pnpm test:content-platform

verify: pnpm --silent proof:admin-content

verify: pnpm test:admin-routes

verify: pnpm test:deploy-targets

verify: pnpm test:security-review

verify: pnpm format:check

verify: git diff --check

verify: node scripts/ci/security-review.mjs --required /tmp/anipotts-com-changed-files.txt
