+++ packages/content/src/admin/index.ts
@@ -1,3 +1,4 @@
 export * from "./content.js";
 export * from "./newsletter.js";
 export * from "./operations.js";
+export * from "./source-content.js";
